Split galaxyManage function out
